refactor: extract the model and app state into src/app/
main.rs 1841 -> 772. This is the last of the four layers to come out, and it
completes the picture the other three were already describing.

  mod.rs      App itself, and impl App
  library.rs  RightView, Section, LibItem, SortMode, Detail, Activated, Library
  action.rs   ActionKind, ActionItem, ActionMenu
  playback.rs NowPlaying, TrackMeta, TrackImage, PlaySource, PlaybackState, seek consts
  state.rs    Services, ThemeState, Transport, Browse/Search/View/Session state
  persist.rs  SavedState, LastPlayed, save_state
  frame.rs    HitRects, FrameOut, ArtRepaint, and the redraw timing policy
  event.rs    engine events and metadata applied to App

The dependency story now reads in one direction and is stated in each module
doc: ui reads &App and writes FrameOut, input mutates App, api talks HTTP and
touches neither, and app is the state in the middle that depends on none of
them. What is left in main.rs is startup, wiring and the event loop.

Pure move. No logic changed - only pub(crate) visibility, use statements, and
rustfmt rewrapping RightView::ALL and the handle_engine_event signature that
pub(crate) pushed past 100 columns. The one deliberate departure from the
api/ and input/ template is that action.rs and library.rs carry no
`use crate::*;`, because they are self-contained and it trips unused_imports
under -D warnings.

SavedState's on-disk format is unchanged: same derives, field names, order and
serde attributes, verified across the whole crate. Only field visibility
changed, which serde does not see.

139 tests still pass; fmt, clippy and test clean with --all-features under
-D warnings.
