+++ src/main.rs
@@ -9,6 +9,10 @@
 /// Lives in the binary (not the library) because it speaks the model types
 /// defined here.
 mod api;
+/// The application state. `ui` reads it, `input` writes it, `api` feeds it over
+/// channels; it depends on none of the three.
+/// Lives in the binary (not the library) because it is what this binary is.
+mod app;
 /// The input layer. Turns terminal and media-key events into `App` mutations
 /// and channel sends — the one layer that writes state.
 /// Lives in the binary (not the library) because it mutates `App`, which is here.
@@ -36,6 +40,7 @@ use ratatui::Frame;
 use ratatui_image::picker::Picker;
 
 use api::*;
+use app::*;
 use input::*;
 use myx::anim::ThemeFade;
 use myx::audio::NUM_BANDS;
@@ -56,843 +61,6 @@ use souvlaki::{
     MediaControlEvent, MediaControls, MediaMetadata, MediaPlayback, MediaPosition, PlatformConfig,
     SeekDirection,
 };
-
-const FADE_MS: u64 = 1500;
-
-/// How far one Shift+arrow press moves the playhead.
-const SEEK_STEP_MS: i64 = 5_000;
-/// Fastest a held Shift+arrow may step. macOS repeats keys ~30×/s; unthrottled,
-/// a one-second hold would throw the playhead 2½ minutes down the track.
-const SEEK_REPEAT: Duration = Duration::from_millis(200);
-/// Quiet time after the last press before the scrub reaches the engine.
-const SEEK_SETTLE: Duration = Duration::from_millis(250);
-
-fn scrub_target(from_ms: u32, duration_ms: u32, delta_ms: i64) -> u32 {
-    (from_ms as i64 + delta_ms).clamp(0, duration_ms as i64) as u32
-}
-
-/// What the album art box owes the next frame.
-///
-/// `ratatui-image` puts the whole image in one cell's symbol and marks the rest
-/// of the box `Skip`, which the diff never touches again — so leftovers stay,
-/// and a re-encode is byte-identical for sixel and iTerm2 and gets discarded.
-/// Blanking the box for one frame is the only change the diff will emit, and it
-/// makes the image that follows one too.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ArtRepaint {
-    /// The box holds what it should.
-    Idle,
-    /// Blank the box this frame.
-    Wipe,
-    /// Draw the art; the wipe has gone out.
-    Draw,
-}
-
-impl ArtRepaint {
-    fn advance(self) -> Self {
-        match self {
-            Self::Wipe => Self::Draw,
-            _ => Self::Idle,
-        }
-    }
-}
-
-/// Ceiling on the redraw rate: one frame per ~60Hz terminal refresh.
-const MIN_FRAME: Duration = Duration::from_millis(16);
-/// Redraw rate while the visualizer or a theme fade is running.
-const ANIM_FRAME: Duration = Duration::from_millis(33);
-/// Redraw rate when nothing changed — enough to keep the clock and progress bar
-/// honest without repainting an identical frame 60 times a second.
-const IDLE_REDRAW: Duration = Duration::from_millis(500);
-/// How often the live queue is re-fetched and the session persisted.
-const SYNC_EVERY: Duration = Duration::from_secs(24);
-
-/// Whether this frame is worth drawing.
-///
-/// Input beats animation beats the idle clock. Smoothness of the recolour comes
-/// from its duration, not from the frame rate: every present makes the terminal
-/// recompose the viewport, and the inline cover shimmers if that happens 60
-/// times a second.
-fn should_draw(dirty: bool, animating: bool, since_last: Duration) -> bool {
-    if dirty {
-        since_last >= MIN_FRAME
-    } else if animating {
-        since_last >= ANIM_FRAME
-    } else {
-        since_last >= IDLE_REDRAW
-    }
-}
-
-// ------------------------------------------------------------------ model
-
-#[derive(Clone, Copy, PartialEq, Eq)]
-enum RightView {
-    NowPlaying,
-    Lyrics,
-    Queue,
-}
-
-impl RightView {
-    const ALL: [RightView; 3] = [RightView::NowPlaying, RightView::Lyrics, RightView::Queue];
-    fn label(self) -> &'static str {
-        match self {
-            RightView::NowPlaying => "Now Playing",
-            RightView::Lyrics => "Lyrics",
-            RightView::Queue => "Queue",
-        }
-    }
-    fn shift(self, delta: isize) -> RightView {
-        let i = RightView::ALL.iter().position(|&v| v == self).unwrap_or(0) as isize;
-        let n = RightView::ALL.len() as isize;
-        RightView::ALL[(i + delta).rem_euclid(n) as usize]
-    }
-}
-
-#[derive(Clone, Copy, PartialEq, Eq)]
-enum Section {
-    Home,
-    Recent,
-    Playlists,
-    Liked,
-    Albums,
-    Artists,
-}
-
-impl Section {
-    const ALL: [Section; 6] = [
-        Section::Home,
-        Section::Liked,
-        Section::Playlists,
-        Section::Albums,
-        Section::Artists,
-        Section::Recent,
-    ];
-    fn label(self) -> &'static str {
-        match self {
-            Section::Home => "Home",
-            Section::Recent => "Recent",
-            Section::Playlists => "Playlists",
-            Section::Liked => "Liked",
-            Section::Albums => "Albums",
-            Section::Artists => "Artists",
-        }
-    }
-    fn index(self) -> usize {
-        Section::ALL.iter().position(|&s| s == self).unwrap_or(0)
-    }
-    fn shift(self, delta: isize) -> Section {
-        let n = Section::ALL.len() as isize;
-        let i = (self.index() as isize + delta).rem_euclid(n) as usize;
-        Section::ALL[i]
-    }
-}
-
-/// A library entry. Behavior on Enter is driven by the flags:
-/// header = non-selectable label; track = play as a track list; play = play this
-/// URI as a context; otherwise = open (drill into) this context.
-#[derive(Clone)]
-struct LibItem {
-    name: String,
-    subtitle: String,
-    uri: String,
-    is_track: bool,
-    is_header: bool,
-    is_play: bool,
-    order: u32, // original fetch position (for the "Added" sort)
-}
-
-impl LibItem {
-    fn track(name: String, subtitle: String, uri: String) -> Self {
-        Self {
-            name,
-            subtitle,
-            uri,
-            is_track: true,
-            is_header: false,
-            is_play: false,
-            order: 0,
-        }
-    }
-    fn ctx(name: String, subtitle: String, uri: String) -> Self {
-        Self {
-            name,
-            subtitle,
-            uri,
-            is_track: false,
-            is_header: false,
-            is_play: false,
-            order: 0,
-        }
-    }
-    fn play(name: String, uri: String) -> Self {
-        Self {
-            name,
-            subtitle: String::new(),
-            uri,
-            is_track: false,
-            is_header: false,
-            is_play: true,
-            order: 0,
-        }
-    }
-    fn header(name: &str) -> Self {
-        Self {
-            name: name.to_string(),
-            subtitle: String::new(),
-            uri: String::new(),
-            is_track: false,
-            is_header: true,
-            is_play: false,
-            order: 0,
-        }
-    }
-}
-
-/// Sort order for browsable lists.
-#[derive(Clone, Copy, PartialEq, Eq)]
-enum SortMode {
-    Added,
-    Title,
-    Artist,
-}
-
-impl SortMode {
-    fn label(self) -> &'static str {
-        match self {
-            SortMode::Added => "added",
-            SortMode::Title => "title",
-            SortMode::Artist => "artist",
-        }
-    }
-    fn next(self) -> SortMode {
-        match self {
-            SortMode::Added => SortMode::Title,
-            SortMode::Title => SortMode::Artist,
-            SortMode::Artist => SortMode::Added,
-        }
-    }
-}
-
-/// Sort a list in place, keeping leading header/play rows pinned at the top.
-fn sort_list(items: &mut [LibItem], mode: SortMode) {
-    let pin = items
-        .iter()
-        .take_while(|i| i.is_header || i.is_play)
-        .count();
-    let tail = &mut items[pin..];
-    match mode {
-        SortMode::Added => tail.sort_by_key(|i| i.order),
-        SortMode::Title => tail.sort_by_key(|i| i.name.to_lowercase()),
-        SortMode::Artist => tail.sort_by_key(|i| i.subtitle.to_lowercase()),
-    }
-}
-
-/// A drill-in detail view (artist / album / playlist contents).
-struct Detail {
-    context_uri: String,
-    title: String,
-    items: Vec<LibItem>,
-    parent_selected: usize,
-}
-
-/// What an action-menu entry does when activated.
-#[derive(Clone)]
-enum ActionKind {
-    ToggleLike {
-        id: String,
-        saved: bool,
-    },
-    Queue {
-        uri: String,
-    },
-    AddToPlaylistMenu {
-        track_uri: String,
-    },
-    AddToPlaylist {
-        playlist_id: String,
-        track_uri: String,
-    },
-    ToggleFollowArtist {
-        id: String,
-        following: bool,
-    },
-    ToggleSaveAlbum {
-        id: String,
-        saved: bool,
-    },
-    FollowPlaylist {
-        id: String,
-    },
-    Play {
-        uri: String,
-        /// Carried so the play path can set `source_name` — without it the
-        /// Queue view's PLAYING FROM header and the persisted resume source
-        /// go stale.
-        name: String,
-    },
-    Open {
-        uri: String,
-        name: String,
-    },
-    CopyLink {
-        uri: String,
-    },
-}
-
-struct ActionItem {
-    label: String,
-    kind: ActionKind,
-}
-
-struct ActionMenu {
-    title: String,
-    items: Vec<ActionItem>,
-    selected: usize,
-}
-
-/// Result of activating (Enter on) a library item.
-enum Activated {
-    None,
-    Open(String, String), // drill into a context (uri, name)
-    Radio(String),        // start this song's radio (seed uri)
-}
-
-#[derive(Default, Clone)]
-struct Library {
-    home: Vec<LibItem>,
-    recent: Vec<LibItem>,
-    playlists: Vec<LibItem>,
-    liked: Vec<LibItem>,
-    albums: Vec<LibItem>,
-    artists: Vec<LibItem>,
-}
-
-impl Library {
-    fn items(&self, s: Section) -> &[LibItem] {
-        match s {
-            Section::Home => &self.home,
-            Section::Recent => &self.recent,
-            Section::Playlists => &self.playlists,
-            Section::Liked => &self.liked,
-            Section::Albums => &self.albums,
-            Section::Artists => &self.artists,
-        }
-    }
-    fn items_mut(&mut self, s: Section) -> &mut Vec<LibItem> {
-        match s {
-            Section::Home => &mut self.home,
-            Section::Recent => &mut self.recent,
-            Section::Playlists => &mut self.playlists,
-            Section::Liked => &mut self.liked,
-            Section::Albums => &mut self.albums,
-            Section::Artists => &mut self.artists,
-        }
-    }
-    fn set(&mut self, s: Section, items: Vec<LibItem>) {
-        match s {
-            Section::Home => self.home = items,
-            Section::Recent => self.recent = items,
-            Section::Playlists => self.playlists = items,
-            Section::Liked => self.liked = items,
-            Section::Albums => self.albums = items,
-            Section::Artists => self.artists = items,
-        }
-    }
-}
-
-struct NowPlaying {
-    uri: String,
-    title: String,
-    artist: String,
-    album: String,
-    duration_ms: u32,
-    position_ms: u32,
-    position_at: Instant,
-    is_playing: bool,
-    cover: Option<Cover>,
-}
-
-struct TrackMeta {
-    uri: String,
-    title: String,
-    artist: String,
-    album: String,
-    duration_ms: u32,
-    image: TrackImage,
-    theme: Option<Theme>,
-}
-
-struct TrackImage {
-    url: Option<String>,
-    image: Option<image::DynamicImage>,
-}
-
-/// What kind of thing is currently playing — persisted so we can resume the real
-/// context (and its live queue) on reboot, not just a bare track.
-#[derive(Clone, Debug, Default, serde::Serialize, serde::Deserialize)]
-enum PlaySource {
-    #[default]
-    None,
-    Context(String), // playlist / album / artist URI
-    Radio(String),   // seed track URI
-    Liked,
-}
-
-/// Persisted across sessions (~/.cache/myx/state.json).
-#[derive(Default, serde::Serialize, serde::Deserialize)]
-struct SavedState {
-    volume: u8,
-    #[serde(default)]
-    shuffle: bool,
-    #[serde(default)]
-    repeat: bool,
-    #[serde(default)]
-    last_played: Option<LastPlayed>,
-    queue: Vec<String>,
-    #[serde(default)]
-    queue_uris: Vec<String>,
-    #[serde(default)]
-    source: PlaySource,
-    #[serde(default)]
-    source_name: String,
-}
-
-#[derive(Default, serde::Serialize, serde::Deserialize)]
-struct LastPlayed {
-    uri: String,
-    title: String,
-    artist: String,
-    album: String,
-    duration_ms: u32,
-    position_ms: u32,
-}
-
-impl SavedState {
-    fn path() -> Option<std::path::PathBuf> {
-        Some(myx::home_dir()?.join(".cache/myx/state.json"))
-    }
-    fn load() -> SavedState {
-        Self::path()
-            .and_then(|p| std::fs::read_to_string(p).ok())
-            .and_then(|s| serde_json::from_str(&s).ok())
-            .unwrap_or_default()
-    }
-    fn save(&self) {
-        let Some(path) = Self::path() else { return };
-        if let Some(dir) = path.parent() {
-            let _ = std::fs::create_dir_all(dir);
-        }
-        if let Ok(json) = serde_json::to_string(self) {
-            let _ = std::fs::write(path, json);
-        }
-    }
-}
-
-/// Mouse hit rects, written by the renderer and read only by `handle_mouse`.
-///
-/// Pure output: every field is (re)set or cleared on each frame that draws the
-/// thing it belongs to, so nothing here is threaded frame-to-frame.
-#[derive(Default)]
-struct HitRects {
-    /// Last-rendered progress-bar rect (for click-to-seek).
-    bar: Option<Rect>,
-    /// Last-rendered sidebar scrollbar track + item count (drag-to-scroll).
-    scroll: Option<Rect>,
-    scroll_len: usize,
-    /// Last-rendered volume-meter bar region (click/drag to set volume).
-    vol: Option<Rect>,
-    /// View tabs in the header.
-    tabs: Vec<(RightView, Rect)>,
-    /// Library list viewport.
-    lib: Option<Rect>,
-}
-
-/// Everything the renderer writes, kept out of `App` so every render function
-/// can take `&App`.
-#[derive(Default)]
-struct FrameOut {
-    hits: HitRects,
-    /// Library viewport start row. Unlike `hits`, this is read-modify-write:
-    /// the renderer feeds the previous frame's value into `scroll_offset` and
-    /// stores the result back, which is what makes scrolling sticky. Owned by
-    /// `run_ui` so it survives across frames.
-    lib_offset: usize,
-}
-
-/// Long-lived services the UI talks to. All three are used through `&self`
-/// (the `Arc<Mutex<_>>` is only ever cloned), so grouping them costs no
-/// borrow flexibility.
-struct Services {
-    engine: Engine,
-    picker: Picker,
-    webapi: Arc<Mutex<WebApi>>,
-}
-
-/// The palette currently on screen, plus the cross-fade walking it towards
-/// the incoming track's palette. `displayed` is what every widget reads;
-/// `target` is only used to snap exactly on completion.
-struct ThemeState {
-    displayed: Theme,
-    target: Theme,
-    fade: Option<ThemeFade>,
-}
-
-impl ThemeState {
-    fn start_fade(&mut self, to: Theme) {
-        self.fade = Some(ThemeFade::new(
-            self.displayed,
-            to,
-            Duration::from_millis(FADE_MS),
-        ));
-        self.target = to;
-    }
-
-    fn advance(&mut self) {
-        if let Some(fade) = &self.fade {
-            self.displayed = fade.current();
-            if fade.is_done() {
-                self.displayed = self.target;
-                self.fade = None;
-            }
-        }
-    }
-}
-
-/// Playback controls and the queue — everything the transport bar and the
-/// persisted `SavedState` care about. None of it touches the playhead.
-struct Transport {
-    shuffle: bool,
-    repeat: bool,
-    volume: u8, // 0..=100 (mirrors the 50% mixer default)
-    queue: Vec<String>,
-    queue_uris: Vec<String>,
-    // Whether real playback has started this session (gates resume-on-play).
-    playback_started: bool,
-    // What's playing (context/radio/liked), for faithful resume on reboot.
-    source: PlaySource,
-    source_name: String,
-}
-
-/// The playhead: what's playing plus the coalesced Shift+arrow scrub state.
-///
-/// These live together because every scrub method touches both — `seek_step`
-/// reads `now.duration_ms` and writes the `seek_*` fields, and
-/// `set_local_position` gates a write to `now` on `seek_target`.
-struct PlaybackState {
-    now: Option<NowPlaying>,
-    // Shift+arrow scrubbing, coalesced (see `seek_step`).
-    seek_target: Option<u32>,
-    seek_last_step: Instant,
-    seek_last_input: Instant,
-}
-
-/// The library browser: what's loaded, where the cursor is, and the drill-in
-/// stack. The viewport offset is not here — it lives in `FrameOut`, since the
-/// renderer owns it across frames.
-struct BrowseState {
-    library: Library,
-    section: Section,
-    selected: usize,
-    sort: SortMode,
-    // Drill-in stack (artist → album → …). Topmost is what's shown.
-    details: Vec<Detail>,
-}
-
-/// The `/` search overlay: whether the prompt is capturing keys, the typed
-/// query, and the results that temporarily replace the library list.
-struct SearchState {
-    input_mode: bool,
-    query: String,
-    searching: bool,
-    search_results: Vec<LibItem>,
-}
-
-/// What the user is looking at: the right pane's mode, the zen (sidebar
-/// hidden) toggle, the lyrics backing the Lyrics view, and the actions
-/// overlay drawn on top of everything.
-struct ViewState {
-    // Which view fills the right pane.
-    mode: RightView,
-    // Sidebar hidden, so the right view (and its cover) gets the whole width.
-    zen: bool,
-    // Lyrics: (timestamp_ms, line). Synced when timestamps are non-zero.
-    lyrics: Vec<(u32, String)>,
-    lyrics_synced: bool,
-    // Context actions menu overlay (opened with `a`).
-    actions: Option<ActionMenu>,
-}
-
-/// Cross-cutting session bookkeeping: what we resumed into, which metadata
-/// fetch is still in flight, and the input timestamps that make Ctrl-C and
-/// double-click work.
-struct SessionState {
-    restore_uri: Option<String>,
-    // Track URI whose metadata was last requested. Fetches run on separate
-    // blocking tasks and can land out of order when skipping quickly, so a
-    // reply for any other track is stale and must be dropped.
-    pending_meta: Option<String>,
-    // Whether we reclaimed a live server-side session (vs. local fallback).
-    reclaimed: bool,
-    // Timestamp of last Ctrl-C — a second press within 1.5s quits.
-    last_ctrl_c: Option<Instant>,
-    last_click: Option<(u16, Instant)>,
-}
-
-struct App {
-    svc: Services,
-    theme: ThemeState,
-    playback: PlaybackState,
-    // Best-effort OS integration. Headless/SSH sessions may not expose the
-    // platform media service, but that must never prevent Myx from playing.
-    media_controls: Option<MediaControls>,
-    status: String,
-    browse: BrowseState,
-    transport: Transport,
-    search: SearchState,
-    view: ViewState,
-    session: SessionState,
-    // What the album art box owes the next frame. See ArtRepaint.
-    art_repaint: ArtRepaint,
-}
-
-fn should_apply_engine_position(from_engine: bool, seek_target: Option<u32>) -> bool {
-    !(from_engine && seek_target.is_some())
-}
-
-impl PlaybackState {
-    fn position_ms(&self) -> u32 {
-        match &self.now {
-            Some(n) if n.is_playing => {
-                (n.position_ms + n.position_at.elapsed().as_millis() as u32).min(n.duration_ms)
-            }
-            Some(n) => n.position_ms.min(n.duration_ms),
-            None => 0,
-        }
-    }
-    /// Move the progress bar, without telling the engine. Reports from the
-    /// engine are ignored mid-scrub — what we painted is newer than anything
-    /// librespot has heard about.
-    fn set_local_position(&mut self, position_ms: u32, from_engine: bool) {
-        if !should_apply_engine_position(from_engine, self.seek_target) {
-            return;
-        }
-        if let Some(n) = self.now.as_mut() {
-            n.position_ms = position_ms.min(n.duration_ms);
-            n.position_at = Instant::now();
-        }
-    }
-    /// Seek to an absolute position (clamped), updating the local display too.
-    fn seek_to(&mut self, engine: &Engine, position_ms: u32) {
-        let Some(dur) = self.now.as_ref().map(|n| n.duration_ms) else {
-            return;
-        };
-        let new = position_ms.min(dur);
-        let _ = engine.seek(new);
-        self.set_local_position(new, false);
-    }
-    /// One Shift+arrow press, moving the playhead by `delta_ms`.
-    ///
-    /// A seek per key repeat overshot the track and made librespot flush and
-    /// refill its audio buffer 30×/s — that pile-up was the stutter. Repeats are
-    /// throttled and the engine seek deferred to `flush_seek`.
-    fn seek_step(&mut self, delta_ms: i64) {
-        let now = Instant::now();
-        if self.seek_target.is_some() && now.duration_since(self.seek_last_step) < SEEK_REPEAT {
-            // The settle timer must see it, or a long hold commits early.
-            self.seek_last_input = now;
-            return;
-        }
-        let Some(dur) = self.now.as_ref().map(|n| n.duration_ms) else {
-            return;
-        };
-        let from = self.seek_target.unwrap_or_else(|| self.position_ms());
-        let target = scrub_target(from, dur, delta_ms);
-        self.seek_target = Some(target);
-        self.seek_last_step = now;
-        self.seek_last_input = now;
-        self.set_local_position(target, false);
-    }
-    /// Commit a finished scrub as a single engine seek, once the keys stop.
-    fn flush_seek(&mut self, engine: &Engine, now: Instant) {
-        if now.duration_since(self.seek_last_input) < SEEK_SETTLE {
-            return;
-        }
-        if let Some(target) = self.seek_target.take() {
-            self.seek_to(engine, target);
-        }
-    }
-}
-
-impl App {
-    fn cur_items(&self) -> &[LibItem] {
-        if let Some(d) = self.browse.details.last() {
-            &d.items
-        } else if self.search.searching {
-            &self.search.search_results
-        } else {
-            self.browse.library.items(self.browse.section)
-        }
-    }
-    fn cur_list_mut(&mut self) -> &mut Vec<LibItem> {
-        if let Some(d) = self.browse.details.last_mut() {
-            &mut d.items
-        } else if self.search.searching {
-            &mut self.search.search_results
-        } else {
-            self.browse.library.items_mut(self.browse.section)
-        }
-    }
-    /// First non-header index (where a fresh selection should land).
-    fn first_selectable(&self) -> usize {
-        self.cur_items()
-            .iter()
-            .position(|i| !i.is_header)
-            .unwrap_or(0)
-    }
-    /// Move the selection by `dir`, skipping header rows, clamped at the ends.
-    fn move_sel(&mut self, dir: isize) {
-        let items = self.cur_items();
-        let n = items.len() as isize;
-        if n == 0 {
-            return;
-        }
-        let mut i = self.browse.selected as isize;
-        loop {
-            i += dir;
-            if i < 0 || i >= n {
-                return;
-            }
-            if !items[i as usize].is_header {
-                self.browse.selected = i as usize;
-                return;
-            }
-        }
-    }
-    /// If the selection landed on a header (e.g. after data loads), bump it off.
-    fn normalize_selection(&mut self) {
-        if self
-            .cur_items()
-            .get(self.browse.selected)
-            .is_some_and(|i| i.is_header)
-        {
-            self.browse.selected = self.first_selectable();
-        }
-    }
-    /// The single entry point for "play this context URI".
-    ///
-    /// Every caller must route through here so `source` / `source_name` stay in
-    /// sync with what is actually playing — they back the Queue view's
-    /// PLAYING FROM header and the resume-on-launch path in `resume_source`.
-    /// `name` is a parameter rather than being derived from `details.last()`
-    /// because the drill-in stack is empty when playing straight from a list.
-    fn play_context_row(&mut self, uri: String, name: String, shuffle: bool) {
-        self.status = format!("starting {name}…");
-        self.transport.source = PlaySource::Context(uri.clone());
-        self.transport.source_name = name;
-        if let Err(e) = self.svc.engine.play_context(uri, shuffle) {
-            self.status = format!("couldn't play: {e:#}");
-        }
-    }
-
-    /// Play whatever's selected (in the current section, or in search results).
-    /// Act on the selected item. Returns what the caller should do next.
-    fn activate(&mut self) -> Activated {
-        let Some(item) = self.cur_items().get(self.browse.selected).cloned() else {
-            return Activated::None;
-        };
-        if item.is_header {
-            return Activated::None;
-        }
-        if item.is_play {
-            // Special synthetic rows: play the Liked list (optionally shuffled).
-            if item.uri == "myx:action:liked-play" {
-                let uris: Vec<String> = self
-                    .browse
-                    .library
-                    .liked
-                    .iter()
-                    .filter(|i| i.is_track)
-                    .map(|i| i.uri.clone())
-                    .collect();
-                if !uris.is_empty() {
-                    self.transport.source = PlaySource::Liked;
-                    self.transport.source_name = "Liked Songs".to_string();
-                    self.status = "starting Liked Songs…".to_string();
-                    // Honour the current shuffle toggle instead of a dedicated row.
-                    if let Err(e) =
-                        self.svc
-                            .engine
-                            .play_tracks(uris, None, 0, self.transport.shuffle)
-                    {
-                        self.status = format!("couldn't play: {e:#}");
-                    }
-                }
-                return Activated::None;
-            }
-            // Inside a drill-in the enclosing title is the better label
-            // ("Chill Vibes"); standalone play rows fall back to their own.
-            let name = self
-                .browse
-                .details
-                .last()
-                .map(|d| d.title.clone())
-                .unwrap_or_else(|| item.name.clone());
-            let shuffle = self.transport.shuffle;
-            self.play_context_row(item.uri, name, shuffle);
-            return Activated::None;
-        }
-        if item.is_track {
-            if self.search.searching {
-                // A search-result song starts that song's radio (seed + similar).
-                self.transport.source = PlaySource::Radio(item.uri.clone());
-                self.transport.source_name = format!("Radio · {}", item.name);
-                return Activated::Radio(item.uri);
-            }
-            // Inside a drill-in → play its context at this track (real queue).
-            if let Some(d) = self.browse.details.last() {
-                let ctx = d.context_uri.clone();
-                self.transport.source = PlaySource::Context(ctx.clone());
-                self.transport.source_name = d.title.clone();
-                self.status = format!("starting {}…", item.name);
-                if let Err(e) = self.svc.engine.play_context_at(
-                    ctx,
-                    Some(item.uri.clone()),
-                    0,
-                    self.transport.shuffle,
-                ) {
-                    self.status = format!("couldn't play: {e:#}");
-                }
-                return Activated::None;
-            }
-            // Section track list.
-            let uris = self
-                .cur_items()
-                .iter()
-                .filter(|i| i.is_track)
-                .map(|i| i.uri.clone())
-                .collect();
-            self.status = format!("starting {}…", item.name);
-            if self.browse.section == Section::Liked {
-                self.transport.source = PlaySource::Liked;
-                self.transport.source_name = "Liked Songs".to_string();
-            } else {
-                self.transport.source = PlaySource::None;
-                self.transport.source_name = self.browse.section.label().to_string();
-            }
-            if let Err(e) =
-                self.svc
-                    .engine
-                    .play_tracks(uris, Some(item.uri.clone()), 0, self.transport.shuffle)
-            {
-                self.status = format!("couldn't play: {e:#}");
-            }
-            return Activated::None;
-        }
-        // Otherwise it's a context (artist / album / playlist) — open it.
-        self.status = format!("opening {}…", item.name);
-        Activated::Open(item.uri, item.name)
-    }
-}
 
 // ------------------------------------------------------------------ main
 
@@ -1571,243 +739,6 @@ fn resume_source(app: &mut App, radio_tx: &flume::Sender<Result<Radio, String>>)
                 }
             }
         }
-    }
-}
-
-/// Does this row carry a playable context URI, and under what name?
-///
-/// Context rows (playlist / album / artist) and the synthesized "▶︎ Play X"
-/// rows both do; headers and tracks do not. Kept pure and free-standing so it
-/// is unit-testable — `App` owns a librespot `Spirc` and cannot be built in a
-/// test. `enter_label` shares this predicate so Enter opens exactly the rows
-/// `P` plays.
-fn context_target(item: &LibItem) -> Option<(String, String)> {
-    (!item.is_header && !item.is_track).then(|| (item.uri.clone(), item.name.clone()))
-}
-
-/// Enter opens context rows and plays everything else.
-fn enter_label(item: Option<&LibItem>) -> &'static str {
-    match item {
-        Some(i) if !i.is_track && !i.is_header => "open",
-        _ => "select",
-    }
-}
-
-/// `P` / `S`: play the highlighted context from anywhere — library section,
-/// search results, or inside a drill-in (`cur_items` resolves all three).
-fn play_selected_context(app: &mut App, shuffle: bool) {
-    let Some(item) = app.cur_items().get(app.browse.selected).cloned() else {
-        return;
-    };
-    match context_target(&item) {
-        Some((uri, name)) => app.play_context_row(uri, name, shuffle),
-        None => app.status = "not a playlist, album, or artist".to_string(),
-    }
-}
-
-/// Snapshot the current session to disk (volume, last track, position, queue).
-fn save_state(app: &App) {
-    let last_played = app.playback.now.as_ref().map(|now| LastPlayed {
-        uri: now.uri.clone(),
-        title: now.title.clone(),
-        artist: now.artist.clone(),
-        album: now.album.clone(),
-        duration_ms: now.duration_ms,
-        position_ms: app.playback.position_ms(),
-    });
-
-    let s = SavedState {
-        volume: app.transport.volume,
-        shuffle: app.transport.shuffle,
-        repeat: app.transport.repeat,
-        last_played,
-        queue: app.transport.queue.clone(),
-        queue_uris: app.transport.queue_uris.clone(),
-        source: app.transport.source.clone(),
-        source_name: app.transport.source_name.clone(),
-    };
-    s.save();
-}
-
-fn handle_engine_event(app: &mut App, ev: EngineEvent, meta_tx: &flume::Sender<TrackMeta>) {
-    // Position ticks would bury everything else in the log.
-    if !matches!(ev, EngineEvent::PositionCorrection { .. }) {
-        liblog(format!("engine: {ev:?}"));
-    }
-    match ev {
-        EngineEvent::TrackChanged { uri } => {
-            app.status = "loading track…".to_string();
-            if let Some(track_id) = track_id_from_uri(&uri) {
-                // Record what we're waiting for so an earlier track's reply,
-                // landing late, is discarded instead of overwriting this one.
-                app.session.pending_meta = Some(format!("spotify:track:{track_id}"));
-                let webapi = app.svc.webapi.clone();
-                let tx = meta_tx.clone();
-                tokio::task::spawn_blocking(move || {
-                    let _ = tx.send(fetch_track_meta(&webapi, &track_id));
-                });
-            }
-        }
-        EngineEvent::Playing { position_ms, .. } => {
-            if !app.transport.playback_started {
-                app.transport.playback_started = true;
-                // Reapply persisted modes + volume to the freshly-started playback.
-                let _ = app.svc.engine.shuffle(app.transport.shuffle);
-                let _ = app.svc.engine.repeat(app.transport.repeat);
-                let _ = app.svc.engine.set_volume(vol_u16(app.transport.volume));
-            }
-            if let Some(n) = app.playback.now.as_mut() {
-                n.is_playing = true;
-            }
-            app.playback.set_local_position(position_ms, true);
-            if let Some(controls) = app.media_controls.as_mut() {
-                let _ = controls.set_playback(MediaPlayback::Playing {
-                    progress: Some(MediaPosition(Duration::from_millis(position_ms as u64))),
-                });
-            }
-        }
-        EngineEvent::Paused { position_ms, .. } => {
-            if let Some(n) = app.playback.now.as_mut() {
-                n.is_playing = false;
-            }
-            app.playback.set_local_position(position_ms, true);
-            if let Some(controls) = app.media_controls.as_mut() {
-                let _ = controls.set_playback(MediaPlayback::Paused {
-                    progress: Some(MediaPosition(Duration::from_millis(position_ms as u64))),
-                });
-            }
-        }
-        EngineEvent::Stopped => {
-            app.playback.now = None;
-            app.transport.playback_started = false;
-            // librespot cleared its context too, so only a fresh load can
-            // resume from here — not a bare `play`.
-            app.session.reclaimed = false;
-
-            if let Some(controls) = app.media_controls.as_mut() {
-                let _ = controls.set_playback(MediaPlayback::Stopped);
-            }
-        }
-        EngineEvent::PositionCorrection { position_ms, .. } => {
-            app.playback.set_local_position(position_ms, true);
-            if let (Some(n), Some(controls)) =
-                (app.playback.now.as_ref(), app.media_controls.as_mut())
-            {
-                let playback = if n.is_playing {
-                    MediaPlayback::Playing {
-                        progress: Some(MediaPosition(Duration::from_millis(position_ms as u64))),
-                    }
-                } else {
-                    MediaPlayback::Paused {
-                        progress: Some(MediaPosition(Duration::from_millis(position_ms as u64))),
-                    }
-                };
-                let _ = controls.set_playback(playback);
-            }
-        }
-        EngineEvent::Reconnecting => {
-            app.status = "connection dropped — reconnecting…".to_string();
-        }
-        EngineEvent::Reconnected => {
-            // The replacement Connect device starts idle, so whatever was
-            // playing is not resumed; say so rather than leave a silent player
-            // looking broken.
-            app.status = if app.transport.playback_started {
-                "reconnected — press play to resume".to_string()
-            } else {
-                "reconnected".to_string()
-            };
-        }
-        EngineEvent::EndOfTrack { .. } => {}
-    }
-}
-
-/// Is this metadata reply the one we are still waiting for?
-///
-/// `None` means nothing specific was requested (e.g. a path that predates the
-/// guard), so accept — the guard only ever discards a reply we can prove is for
-/// a different track.
-fn meta_is_current(pending: Option<&str>, meta_uri: &str) -> bool {
-    pending.is_none_or(|p| p == meta_uri)
-}
-
-fn apply_meta(
-    app: &mut App,
-    meta: TrackMeta,
-    lyrics_tx: &flume::Sender<(Vec<(u32, String)>, bool)>,
-) {
-    // Metadata fetches run on independent blocking tasks, so skipping quickly
-    // (n/b) can land an earlier track's reply after a later one. Applying it
-    // would replace the whole NowPlaying — title, artist and cover — with the
-    // wrong track's data.
-    if !meta_is_current(app.session.pending_meta.as_deref(), &meta.uri) {
-        return;
-    }
-
-    let cover = meta
-        .image
-        .image
-        .as_ref()
-        .map(|img| Cover::from_image(img.clone(), app.svc.picker.clone()));
-    // A different cover encodes to a different symbol, so the diff emits it on
-    // its own — no wipe, which would flash a blank box between the two covers.
-    app.art_repaint = ArtRepaint::Draw;
-    app.status.clear();
-    app.view.lyrics.clear();
-    app.view.lyrics_synced = false;
-
-    // Fetch synced lyrics from lrclib for the new track.
-    if !meta.title.is_empty() {
-        let (artist, title, album, dur) = (
-            meta.artist.clone(),
-            meta.title.clone(),
-            meta.album.clone(),
-            meta.duration_ms,
-        );
-        let tx = lyrics_tx.clone();
-        tokio::task::spawn_blocking(move || {
-            let _ = tx.send(fetch_lyrics_blocking(&artist, &title, &album, dur));
-        });
-    }
-
-    app.playback.now = Some(NowPlaying {
-        uri: meta.uri,
-        title: meta.title,
-        artist: meta.artist,
-        album: meta.album,
-        duration_ms: meta.duration_ms,
-        position_ms: app
-            .playback
-            .now
-            .as_ref()
-            .map(|n| n.position_ms)
-            .unwrap_or(0),
-        position_at: Instant::now(),
-        is_playing: app
-            .playback
-            .now
-            .as_ref()
-            .map(|n| n.is_playing)
-            .unwrap_or(app.transport.playback_started),
-        cover,
-    });
-
-    if let Some(theme) = meta.theme {
-        app.theme.start_fade(theme);
-    }
-
-    if let Some(controls) = app.media_controls.as_mut() {
-        let _ = controls.set_metadata(MediaMetadata {
-            title: app.playback.now.as_ref().map(|n| n.title.as_str()),
-            artist: app.playback.now.as_ref().map(|n| n.artist.as_str()),
-            album: app.playback.now.as_ref().map(|n| n.album.as_str()),
-            cover_url: meta.image.url.as_deref(),
-            duration: app
-                .playback
-                .now
-                .as_ref()
-                .map(|n| Duration::from_millis(n.duration_ms as u64)),
-        });
     }
 }
